--- data/spawner.ron
@@ -1,4 +1,0 @@
-(
-	initial_duration: 5.0,
-	formation_pool: Easy,
-)
