+++ src/arena/mod.rs
@@ -3,6 +3,6 @@ mod barrier;
 mod gate;
 
 pub use self::{
-    barrier::spawn_barriers_system,
+    barrier::{spawn_barriers_system, ArenaBarrierComponent},
     gate::{despawn_gates_system, spawn_despawn_gates_system, EnemyReachedBottomGateEvent},
 };
