+++ src/run/formation.rs
@@ -1,14 +1,12 @@
-use std::collections::HashMap;
-
-use crate::{
-    game::GameParametersResource,
-    spawnable::{spawn_mob, FormationPool, MobsResource, SpawnableType},
-};
 use bevy::prelude::*;
 use bevy_rapier2d::physics::RapierConfiguration;
 use serde::Deserialize;
+use std::collections::HashMap;
+
+use crate::{game, spawnable};
 
 pub type FormationPoolsResource = HashMap<FormationPoolType, FormationPool>;
+pub type FormationPool = Vec<Formation>;
 
 #[derive(Deserialize, Debug, Hash, PartialEq, Eq, Clone)]
 pub enum FormationPoolType {
@@ -20,7 +18,7 @@ pub enum FormationPoolType {
 #[derive(Deserialize, Clone)]
 pub struct FormationSpawnable {
     /// Type of spawnable in formation
-    pub spawnable_type: SpawnableType,
+    pub spawnable_type: spawnable::SpawnableType,
     /// Position of the spawnable
     pub position: Vec2,
 }
@@ -40,16 +38,16 @@ impl Formation {
     /// Spawn all spawnables in the formation at once
     pub fn spawn_formation(
         &self,
-        mobs: &MobsResource,
+        mobs: &spawnable::MobsResource,
         commands: &mut Commands,
         rapier_config: &RapierConfiguration,
-        game_parameters: &GameParametersResource,
+        game_parameters: &game::GameParametersResource,
     ) {
         for formation_spawnable in self.formation_spawnables.iter() {
             // TODO: add cases for items, consumables, etc, as they are added
             // spawn enemy
             match &formation_spawnable.spawnable_type {
-                SpawnableType::Mob(mob_type) => spawn_mob(
+                spawnable::SpawnableType::Mob(mob_type) => spawnable::spawn_mob(
                     mob_type,
                     mobs,
                     formation_spawnable.position,
@@ -60,5 +58,26 @@ impl Formation {
                 _ => {}
             }
         }
+    }
+}
+
+/// Event for spawning formations
+pub struct SpawnFormationEvent {
+    //pub formation_pool: FormationPoolType,
+    pub formation: Formation,
+}
+
+/// Manages spawning of formations
+pub fn spawn_formation_system(
+    mut commands: Commands,
+    mut event_reader: EventReader<SpawnFormationEvent>,
+    mobs: Res<spawnable::MobsResource>,
+    rapier_config: Res<RapierConfiguration>,
+    game_parameters: Res<game::GameParametersResource>,
+) {
+    for event in event_reader.iter() {
+        event
+            .formation
+            .spawn_formation(&mobs, &mut commands, &rapier_config, &game_parameters);
     }
 }
