--- src/spawnable/spawn/mod.rs
@@ -1,7 +1,0 @@
-mod formation;
-mod spawner;
-
-pub use self::formation::{Formation, FormationPoolType, FormationPoolsResource};
-pub use self::spawner::{
-    spawn_formation_system, FormationPool, SpawnerResource, SpawnerResourceData,
-};
